Changed call_class type to be long long in order to be able to add more #pragma aux parameters. Added "rdosdev" aux in order to be able to generate the special prolog sequence required.
[git-p4: depot-paths = "//depot/openwatcom/": change = 37123]

+++ bld/cg/h/cgaux.h
@@ -93,7 +93,7 @@ typedef enum  {
     #error too many attributes
 #endif
 
-typedef unsigned long call_class;
+typedef unsigned long long call_class;
 typedef unsigned long byte_seq_len;
 
 #include "cgauxa.h"
